refactor(result-dashboard): DashboardView を抽出し §10.2 3 層分離を本実装化
- 描画専用の `DashboardView` を新規追加（副作用ゼロ・props ベースの pure component、`useMemo` のみ許容）
- `ResultDashboard` は `useCountUp` × 4 / `useMediaQuery` / PDF 制御のコンテナ責務に純化
- `animated` / `isMobile` などの境界値はコンテナで導出し props 注入することで描画層を hook フリーに保つ
- 隠しマウント `PdfDashboard` JSX は `pdfDashboardRef` / `generatedAt` / `inputs` 依存のためコンテナ側に残置
- JSDoc から「暫定構造」「別 Issue として申し送り」記述を更新し 3 層分離の現状を反映
- `PdfDashboard.tsx` は本 PR では編集せず、仕様書 §10.4 に準じ独立 JSX を維持

Closes #73
